+++ tui/entry.go
@@ -1,18 +1,16 @@
 package tui
 
 import (
-	. "togo/tui/constants"
-
 	tea "github.com/charmbracelet/bubbletea"
 )
 
-func (model Model) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
+func (model Model) EntryUpdate(msg tea.Msg) (tea.Model, tea.Cmd) {
 	var cmd tea.Cmd
 
-	model.textInput, cmd = model.textInput.Update(msg)
+	model.TextInput, cmd = model.TextInput.Update(msg)
 	return model, cmd
 }
 
-func (model Model) View() string {
-
+func (model Model) EntryView() string {
+	return ""
 }
