+++ tui/task.go
@@ -1,15 +1,13 @@
 package tui
 
 import (
-	. "main"
-
 	tea "github.com/charmbracelet/bubbletea"
 )
 
-func (model Model) Update(msg tea.Msg) (tea.Model, tea.Cmd) {
+func (model Model) TaskUpdate(msg tea.Msg) (tea.Model, tea.Cmd) {
 	return model, nil
 }
 
-func (model Model) View() string {
-
+func (model Model) TaskView() string {
+	return ""
 }
